Yoink auth into the feed generator package for a tiny refactor for now
